add using MainTitlt and Subtitle to Categories, rename CategoriesItem to CategoriesCard
