+++ src/components/CategoriesCard/CategoriesCard.module.scss
